requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-spectacular==0.13.2
importlib-metadata==3.7.0
inflection==0.5.1
jsonschema==3.2.0
Markdown==3.3.4
PyJWT==1.7.1
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
zipp==3.4.0
